Changed PricingInfo.amount from int to float
git-svn-id: https://svn.code.sf.net/p/osp-toolkit/code/main@390 43700839-f8e2-0310-9689-8ed4146226e7

+++ include/osp/ospdatatypes.h
@@ -208,7 +208,7 @@ typedef enum
 
 typedef struct
 {
-    unsigned amount;
+    float amount;
     unsigned increment;
     unsigned char unit[OSPC_UNITSIZE];
     unsigned char currency[OSPC_CURRENCYSIZE];
